move fetch data function to server path and rename it
